requirements.txt
Django==1.11.15
django-tastypie==0.14.2
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-decouple==3.1
python-mimeparse==1.6.0
pytz==2018.7
six==1.11.0
